backend/requirements.txt
# 装修决策Agent - Python依赖包

# FastAPI框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

SQLAlchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.0

# 缓存
redis==5.0.1
hiredis==2.2.3

# 微信SDK
wechatpy==1.8.18
cryptography>=41.0.0

# HTTP客户端
httpx==0.25.2
aiofiles==23.2.1

# 阿里云OCR
alibabacloud-tea-openapi==0.3.9
alibabacloud-ocr-api20210707==2.0.0

# 文件存储
oss2==2.18.4

# 工具库
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
python-multipart==0.0.6

# 日志
loguru==0.7.2

# 环境变量
python-dotenv==1.0.0

# DeepSeek API客户端
openai==1.3.7

# 安全和限流
slowapi==0.1.9
tenacity==8.2.3
circuitbreaker==2.0.0

# 微信支付（已存在，保留）
wechatpy==1.8.18
